Extract flowchartFromSpans. Add sequenceDiagramFromSpans
